+++ src/main/java/org/prgrms/springbootbasic/repository/VoucherRepository.java
@@ -1,9 +1,9 @@
 package org.prgrms.springbootbasic.repository;
 
 
-import org.prgrms.springbootbasic.voucher.Voucher;
+import org.prgrms.springbootbasic.entity.voucher.Voucher;
 
-import java.util.Map;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -14,5 +14,5 @@ public interface VoucherRepository {
 
     Optional<Voucher> findById(UUID VoucherId);
 
-    Map<UUID, Voucher> findAll();
+    List<Voucher> findAll();
 }
